A lot of things new, read more below:
[*] Added run_as_daemon feature.
[*] Separating get_time() from battery_monitor(): If another function
wants to use it, perfect!
[*] Logging when closing the GUI.
[*] Using stdint integers instead of generic int, size_t in some places.

+++ gui.h
@@ -3,11 +3,12 @@
 
 #include <gtk/gtk.h>
 #include <stdbool.h>
+#include <stdint.h>
 
 #include "format.h"
 #include "logger.h"
 #include "power_options.h"
 
-int gui(int* argc, char **argv);
+uint8_t gui(int* argc, char **argv);
 
 #endif
